[lib] Improve Farcaster sync loading state
Summary:
The sync may take a lot of time for power users, so I checked what can be done in an hour with the AI in order to make the progress more predictable.

{F12989129}

{F12989132}

Depends on D15306

Test Plan: Checked if the loading state is displayed correctly on native and web.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15309