[blob] Store media info in DDB
Summary:
For multimedia blobs, we want to store some metadata, mainly content-type which is required for clients to properly parse media payload.
Additionally, added some custom metadata string, without any hard constraints (it's up to the client). It can be used to store some metadata that Comm supports but Farcaster not.

Multimedia blobs are distinguished by presence of this attribute in DB.

Depends on D15315

Test Plan: Tested on staging

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15316