[blob] Add endpoint to fetch media
Summary:
Endpoint that is going to be called by both Comm and Farcaster clients to download and view multimedia. Thus it has to be unauthenticated.
It allows downloading only blobs that are multimedia blobs (have the media info attribute). Also, it takes care of setting correct response content type.

Depends on D15317

Test Plan: Tested on staging and in later diffs. Verified that only multimedia blobs are reachable by it

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15318