[lib] Fetch all the inbox types
Summary:
We should additionally fetch archived and request inbox categories.

https://linear.app/comm/issue/ENG-11189/handle-conversation-requests

Depends on D15311

Test Plan: Checked if the requests also were fetched. The problem with them is that fetching messages fails for them because of `Not authorized to access this conversation` error. It is expected and we should handle fetching the messages again after joining a thread https://linear.app/comm/issue/ENG-11186/fetch-all-the-messages-when-a-user-is-added-to-a-thread.

Reviewers: kamil, bartek