[lib] Create a new message type for compound reactions
Summary:
Farcaster reactions are delivered as properties of a message. Also, they don't say who reacted. This means we can't use the existing reaction type - we can't deterministically create a set of react and unreact messages based on a received message.

The solution to that is a new message type. For each message there is going to be at most one such a message. We can deterministically create it from the Farcaster data.

https://linear.app/comm/issue/ENG-11244/support-receiving-reactions-on-comm

Depends on D15333

Test Plan: Tested with the rest of the stack by fetching Farcaster messages and verifying that all the reactions are shown.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15334