[lib] Load Farcaster messages while scrolling
Summary:
It is possible that a message changes on Farcaster between saving it in the DB and showing it on the screen. This happens e.g. when someone reacts to it. And we don't receive any update on the websocket when it happens.
The solution is to fetch messages both from Farcaster and DB while scrolling the screen. By doing this we ensure that the correct message content is being shown. If a message is no longer available on Farcaster, it will still appear on the screen.

https://linear.app/comm/issue/ENG-11245/load-farcaster-messages-while-scrolling

Depends on D15374

Test Plan:
Create a thread with a lot of messages. React to it on Farcaster and open the thread on Comm. Scroll to the message - the reaction should be shown.

Reviewers: kamil, bartek