[native] Keep async/await syntax in the Farcaster signer bundle
Summary:
In the release mode, the signing didn't work because the `regeneratorRuntime` wasn't present. This runtime was required to support `async/await` syntax in older JS versions. We can target more modern browsers so that the `async/await` syntax can be kept, which means the runtime isn't required.

https://linear.app/comm/issue/ENG-11261/connecting-dcs-never-finishes

Test Plan:
Run the iOS app in release mode and confirm that connecting DCs work.

Reviewers: kamil, bartek