[lib] Fix fetching Farcaster conversations after connection
Summary:
We're using a flag, stored in sync metadata store, to check if we should sync the conversations. This flag can have three values - undefined is returned when the backup wasn't restored yet. We were setting this flag to false when we were unlinking the account. But we didn't set it during the connection - which blocked the sync screen from appearing.

The solution is to set this flag to false every time we set the connection.

https://linear.app/comm/issue/ENG-11270/nothing-happens-after-connecting-farcaster-dcs

Test Plan:
Register a new account connected with Farcaster but not DCs. Connect the DCs - the loading screen should appear.

Reviewers: kamil, bartek