[lib] Fix the fetching again
Summary:
A solution from D15359 has a problem - every time the app is opened, we fetch the threads again. This comes from the fact that in `FarcasterDataHandler` we're setting the local DCs support on each app open - so we can't modify the flag there.

The correct solution is to set the flag to false when we connect. This solves the issue on the connecting device.
All the other devices aren't affected because they will receive a `FARCASTER_CONNECTION_UPDATED` message where we set the flag to false.
I'm not sure about an edge case where a new secondary device was added before the primary device uploaded a log where the flag is set - this is a more general concern - created https://linear.app/comm/issue/ENG-11280/is-it-possible-to-miss-some-data-in-the-backup to discuss.

https://linear.app/comm/issue/ENG-11270/nothing-happens-after-connecting-farcaster-dcs

Depends on D15360

Test Plan:
Register a new account connected with Farcaster but not DCs. Connect the DCs - the loading screen should appear.
Close and reopen the app - the loading screen should not appear.
Also checked the secondary device
- Logged in after the connection - the conversations were there and the loading screen didn't appear.
- Refreshed the page - the conversations were still there and loading screen didn't appear.
Disconnected and connected - loading screens appeared on both devices.

Reviewers: kamil, bartek