[native] Bump @redux-devtools/cli to 4.0.3
Summary:
Earlier on I thought this would be necessary to address [ENG-11237](https://linear.app/comm/issue/ENG-11237/all-the-android-buildkite-jobs-started-failing). It turns out that it's not, but it's still a good idea to update.

We can't use 4.0.4 since it bumps to React 19, which causes [this issue](https://github.com/reduxjs/redux-devtools/issues/1931). I had to remove our React 19 Yarn resolutions since they were causing the same issue, which caused Redux Dev Tools to be broken (even before this diff). The React 19 Yarn resolutions were something I tried when investigating a solution to [ENG-11240](https://linear.app/comm/issue/ENG-11240/markdown-broken-on-web-with-react-19), but they did not turn out to resolve the issue and were not actually necessary, as confirmed by testing in the Test Plan below.

Depends on D15362

Test Plan:
1. Test Redux Dev Tools on native and make sure it works again
2. Test the removal of the React 19 Yarn resolution on my personal keyserver running on production. Make sure we don't reintroduce "Minified React error #525" (see [ENG-11240](https://linear.app/comm/issue/ENG-11240/markdown-broken-on-web-with-react-19))

Reviewers: tomek, bartek

Differential Revision: https://phab.comm.dev/D15363