[lib][native][web] Track connecting Farcaster DC alert separate from connecting Farcaster alert
Summary:
We currently use the same `alertTypes` constant for both of these. This has a risk that somebody would see the "connect Farcaster" alert, choose to enable it, and then never see the "connect Farcaster DC" alert because `shouldSkipConnectFarcasterAlert` filters after you see the alert once.

To add this new `alertTypes` constant we need to make sure it's present in the Redux store on `native`. This diff introduces a no-op migration on web to keep the store version consistent between platforms.

Test Plan: Careful code inspection and Flow

Reviewers: tomek, kamil

Differential Revision: https://phab.comm.dev/D15387