[lib] Remove invariant on media_infos in rawMessageInfoFromClientDB
Summary:
This invariant can lead to [ENG-11326](https://linear.app/comm/issue/ENG-11326/media-infos-must-be-defined-crash-on-web) and [ENG-9079](https://linear.app/comm/issue/ENG-9079/media-uploads-can-be-partially-assigned).

The code below can already handle the case where it is not set. While it's true that we should eventually investigate what is causing this, for now it's better to avoid crashing the app.

Test Plan: Flow

Reviewers: tomek, kamil

Subscribers: