[native] fix uploading compaction after restoring backup
Summary:
[ENG-11311](https://linear.app/comm/issue/ENG-11311/fix-uploading-compaction-after-restoring-backup).

It was easier than I initially thought; not sure why this condition was missing initially.

This solution is not ideal, there is still minor chance for edge-case where logs are uploaded for the previous backup, there are three ways of mitigating this, created a follow-up task: []().

Test Plan:
Things I tested:
1. Restoring primary after logout
2. Restoring primary after deleting app
3. Adding secondary device
4. Checking of logs upload are starting from id 0 (after uploading compaction)
5. Killed the app after restore but before uploading compaction, then opened app again and checked if compaction is uploaded

Reviewers: tomek