[lib] reset `user_data_backup_started` state when starting app
Summary:
[ENG-11313](https://linear.app/comm/issue/ENG-11313/investigate-why-ashoat-had-no-compactions-between-september-1st-and).

Handling this in reducer because it is easier ot avoid edge-cases, in `BackupHandlerContextProvider` it would be more a lot more complex to implement this correct (this state is used to prevent running upload in parallel).

Test Plan: Explained in [here](https://linear.app/comm/issue/ENG-11313/investigate-why-ashoat-had-no-compactions-between-september-1st-and#comment-60a0a331), point 2.

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15395