[lib] Run the lightweight sync
Summary:
Run the sync when necessary

https://linear.app/comm/issue/ENG-11209/only-refresh-farcaster-threads-that-are-visible-on-screen

Depends on D15407

Test Plan:
On native, close and open the app and check if the sync was performed.
On web, refresh the page and check if the sync was performed.
Added a secondary device and checked if the sync was NOT performed after the restoration.
On Farcaster, added a thread, edited the name and the description of the second thread, and left the third one. Then checked if the sync handles all these cases correctly.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15408