[lib] Add an option to refresh a thread
Summary:
On both platforms, add an option to refresh a Farcaster thread

https://linear.app/comm/issue/ENG-11348/add-option-to-refresh-farcaster-conversations

{F15257171}

{F15257173}

Test Plan: Checked if refreshing works by updating a name on Farcaster and clicking the button in Comm. Checked if the buttons are displayed only for Farcaster. Also checked if it isn't possible to click the button during the refresh.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15410