[lib] Add an option to refresh a thread
Summary:
On both platforms, add an option to refresh a Farcaster thread

https://linear.app/comm/issue/ENG-11348/add-option-to-refresh-farcaster-conversations

Test Plan:
Checked if refreshing works by updating a name on Farcaster and clicking the button in Comm. Checked if the buttons are displayed only for Farcaster. Also checked if it isn't possible to click the button during the refresh.

Reviewers: kamil