[lib] Limit the batch size when syncing Farcaster threads
Summary:
When a batch has a lot of messages, split it into multiple redux dispatches. In this solution, we still don't split a conversation into multiple batches - it is only a problem if a conversation has a lot of messages (a couple MB of text).

https://linear.app/comm/issue/ENG-11360/reduce-the-size-of-batches-when-syncing-farcaster-data

Test Plan: Set a low batch size limit and check if the batches were dispatched separately - based on the logs

Reviewers: kamil