[tunnelbroker][lib] Refresh Farcaster unread statuses
Summary:
When the number of unseen threads changes, a WebSocket message is sent. When we receive such a message, we fetch the inbox on Tunnelbroker and compute the unread statuses - then we deliver the statuses to clients who update the threads accordingly.

https://linear.app/comm/issue/ENG-11297/unread-status-is-not-synced-between-farcaster-and-comm-apps

Test Plan: Open the Comm app, and see if:
1. Opening an unread thread on Farcaster marks it as read on Comm
2. Sending a Farcaster message to a user, marks the thread on the user's device as unread

Reviewers: kamil