[lib] Don't remove threads if fetching inboxes fails
Summary:
We have a mechanism that removes threads that were deleted on Farcaster during search. We fetch the whole inbox, compare it with the thread store, and delete threads that aren't present in the inbox. We should do that only if fetching the inbox succeeds.

https://linear.app/comm/issue/ENG-11394/when-fetching-the-inbox-fails-we-remove-all-the-farcaster-messages

Test Plan: Make sure fetching inboxes fails and check if the threads persist after opening the app.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15430