[tunnelbroker][lib][native][web] Handle Farcaster badge count notifs
Summary:
When we see a Farcaster message about unseen count, we're sending notifications so that clients can update their badges. We're using the same storage mechanism as for the keyservers - this is a small shortcut.

https://linear.app/comm/issue/ENG-11384/unread-count-on-inbox-icon-in-app-doesnt-match-badge-count-on-ios

Test Plan:
Currently tested only on iOS, but will test on other platforms as well:
When the app is opened, or in the background, or killed: open a thread in the Farcaster app and see if the badge count displayed on the app icon is updated consistently with the in-app badge count.

Reviewers: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15441