[tunnelbroker] Don't send notifs from requests
Summary:
We shouldn't send notifs when a new message in a request conversation appears.

https://linear.app/comm/issue/ENG-11412/notif-received-for-farcaster-request

Test Plan: Requests for me are broken and I can't see messages in them, but as a workaround I tested sending a message in a normal conversation and checked that the category is correctly parsed. The rest doesn't need to be tested, because returning `None` is already happening in other branches.

Reviewers: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15449