[tunnelbroker] Don't persist inbox updated messages
Summary:
When a client is offline, we should avoid queueing such messages because their processing is expensive. After the app is opened, we're performing a lightweight sync regardless. The solution is to stop persisting the messages - when a connection is open, they are delivered. If it is closed, they won't be delivered after the app is opened.

https://linear.app/comm/issue/ENG-11454/replace-older-farcasterinboxstatus-messages-when-generating-new-one

Test Plan: Open the app, read a thread on Farcaster and see if a message is delivered. Close the app, read another thread, and open the app - the message isn't delivered.

Reviewers: kamil