[tunnelbroker] use proper `device_id` when deleting message after AMQP error
Summary: One thing I discovered when looking at D15484. Not that this is only on error handling, so on things like RabbitMQ problems, and this code shouldn't be executed a lot so we shouldn't worry that this caused persistence ofa  huge amount of non-deleted messages.

Test Plan: Ask Tomek to run this code

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15485