[lib] Introduce the option to cancel Farcaster sync
Summary:
Add an option to cancel, which hides the loading screen and stops the processing gracefully.

{F16390423}

https://linear.app/comm/issue/ENG-11455/add-cancel-button-to-farcaster-conversation-sync-on-native

Test Plan:
Check if canceling works and stops the sync.
Check if opening the app after the cancel doesn't cause the lightweight sync to fetch all the messages.
Check if receiving the refresh inbox messages doesn't cause the whole threads to be fetched.
Log in on the web and check if the sync also doesn't happen there (users can manually sync from the settings screen).
Check if clicking the sync button starts the sync again.
All the checking was performed by looking at the Tunnelbroker logs to see if there were requests about threads / messages.

Reviewers: kamil, ashoat

Reviewed By: kamil, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15487