[native] Imperatively refresh lifecycleState
Summary:
This diff attempts to address [ENG-11465](https://linear.app/comm/issue/ENG-11465/backup-handler-stopped-after-restore). It adds two places where we call `onLifecycleChange(getCurrentLifecycleState())` to imperatively refresh the `lifecycleState`:

1. When the `LifecycleHandler` first renders. This handles a case where the app is backgrounded before `PersistGate` first renders `LifecycleHandler`.
2. When backup restore concludes. This handles the main case observed in the linked task, where "thrashing" during backup restore results in an incorrect `lifecycleState` after the restore completes.

Test Plan: I tried restoring several times with this change and did not observe the issue. However, the issue was hard to reproduce... I only ever succeeded 4 or 5 times in total. So it's hard to say if it's solved. That said, I think the changes here are safe and worth making.

Reviewers: tomek, kamil

Differential Revision: https://phab.comm.dev/D15491