[lib] Support pinning messages
Summary:
Support the pinning. It requires introducing a new property to threads where the pinned message IDs are stored. Handling the C++ side is placed later in the stack.

https://linear.app/comm/issue/ENG-11220/handle-message-pinning

Test Plan: Tested in combination with the rest of the stack: both on native and web, test if pinning and unpinning works. Also tested if the pins persist app reopens. Checked if the pins are correctly saved on Farcaster.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15498