[lib] Move pinning messages to a spec
Summary:
Create a function in a spec and use it.

https://linear.app/comm/issue/ENG-11220/handle-message-pinning

Depends on D15498

Test Plan: Tested in combination with the rest of the stack: both on native and web, test if pinning and unpinning works. Also tested if the pins persist app reopens. Checked if the pins are correctly saved on Farcaster.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15499