[lib] Move fetching pinned messages to a spec
Summary:
Move to a spec and use the spec instead.

https://linear.app/comm/issue/ENG-11220/handle-message-pinning

Depends on D15499

Test Plan:
Tested in combination with the rest of the stack: both on native and web, test if pinning and unpinning works. Also tested if the pins persist app reopens. Checked if the pins are correctly saved on Farcaster.

Reviewers: kamil