[terraform] migrate alarms to Fargate instances on production
Summary:
[ENG-11520](https://linear.app/comm/issue/ENG-11520/terraform-error-with-aws-ami).

The issue was that the AMI query was looking for a specific version (amzn2-ami-ecs-hvm-2.0.20231024-x86_64-ebs) that no longer exists. I changed it to use a wildcard
  pattern (amzn2-ami-ecs-hvm-*-x86_64-ebs) which will match the most recent ECS-optimized Amazon Linux 2 AMI.

Additionally, we're going to shut it down in couple of days anyway, so I added `ignore_changes`, to the launch template. This will prevent Terraform from updating the ECS instances when the AMI changes - there is no point in that.

Without this diff, terraform deployments are blocked.

Depends on D15519

Test Plan: Terraform plan

Reviewers: tomek, ashoat