[lib] Create proper Farcaster protocol replies
Summary:
When a user starts the reply experience, we're saving this in a state. When sending a message, if it still looks like a reply, we remove the reply prefix and send a request with the reply.

https://linear.app/comm/issue/ENG-11289/farcaster-create-a-protocol-reply

Test Plan:
On native tested reply action from a tooltip and from a swipe gesture. On web the tooltip action.
Tested replying with some text and checked in the Farcaster app if the result is correct.
Checked if modifying the target message text doesn't result in a reply being created - it is a text message with some markdown.
Checked if we correctly leave the reply state after sending a message. Created a reply, copied its content and sent it again - the second message contains the markdown instead of the reply.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15526