[lib] Upload the Farcaster thread avatars
Summary:
Upload the avatar and pass the result to the API.

https://linear.app/comm/issue/ENG-11176/implement-updating-thread-avatar

Depends on D15533

Test Plan:
Check if it is possible to update the avatar by using a camera or a gallery on native, and by selecting an image on web.
Check removing the avatar.
Confirm that the avatar is correctly updated in the Farcaster app. (Note: when testing it on web, I was using my local setup, which means that I didn't fully test the flow - the avatar in the Farcaster app was blank. On my local apps it worked fine, though, and I'm using the same logic as for multimedia messages, so it should be ok.)

Reviewers: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15534