[lib][web][native] introduce `olmSupportRequired` and create session with based on code version
Summary:
[ENG-11545](https://linear.app/comm/issue/ENG-11545/handle-olm-x3dh-bug).

Assuming keyserver is going to be deployed first, and we have only one keyserver so by default, hardcoding `olmSupportRequired: false` for keyservers.

This should be deprecated and in the future, each device should use vodozemac only.

Depends on D15550

Test Plan: Test creating session when this return false (by default it was true).

Reviewers: ashoat