[lib][web][keyserver] add Vodozemac library and `.wasm` file handling (#324)
* [lib][web][keyserver] add Vodozemac library and `.wasm` file handling

Summary:
[ENG-11531](https://linear.app/comm/issue/ENG-11531/migrate-the-full-crypto-api-to-vodozemac).

Using the same pattern as have for existing `.wasm` files (Olm, SQLite and Backup Client).

Currently linking my local library, so CI will fail. This should be updated to NPM package after https://github.com/CommE2E/vodozemac/pull/3 is merged and published.

This diff shouldn't be landed without following one with API migration, because it replaces initializing Olm with Vodozemac, so Olm is not working anymore.

Depends on D15546

Test Plan: Run web and keyserver, vodozemac should be available, usage tested in next diff

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D15547

* [lib][web][keyserver] replace Olm with Vodozemac's Olm (#325)

* [lib][web][keyserver] replace Olm with Vodozemac's Olm

Summary:
[ENG-11531](https://linear.app/comm/issue/ENG-11531/migrate-the-full-crypto-api-to-vodozemac).

There is one thing with naming which is not obvious, this is using Olm name - Olm is still part of Vodozemac crate, so I think this is still correct, but I am okay with updating this too.

I tried to make API as similar as it is possible, so there are files when changing import is everything we have to do.

I redefined `EncryptResult` in `lib/`. Importing it from Olm was bad idea because it was used on native too, which causes Olm/Vodozemac wasm to be bundled into native, which worked with Olm but not with Vodozemac.

Depends on D15547

Test Plan: Test encrypt/decrypt and creating session with Olm and Vodozemac.

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D15548

* [native] implement CXX bindings to Vodozemac (#326)

Summary:
[ENG-11531](https://linear.app/comm/issue/ENG-11531/migrate-the-full-crypto-api-to-vodozemac).

Making vodozemac available on native.

Branch needs to be updated after PRs are merged

Test Plan: Tested in next diff

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D15549